refactor(store-core): move permission-mode handlers into handlers/permission.ts (audit P2-3)
Move the permission-mode handlers (handlePermissionModeChanged,
handleAvailablePermissionModes + PermissionMode, handleConfirmPermissionMode +
PendingPermissionConfirm) out of handlers/index.ts into handlers/permission.ts,
joining the permission-request handlers already there. Re-exported via the
existing export * from './permission' barrel line, so the public surface is
unchanged.

handleAuthOk folds the available-permission-modes list into its payload, so
index.ts imports handleAvailablePermissionModes + PermissionMode back from
./permission. No cycle: ./permission only imports ../types, ../utils, ./_shared
— never ./index. This unblocks extracting the auth family next (auth.ts will
import the same two symbols from ./permission instead).

Pure move — both moved sub-ranges byte-identical to origin/main. index.ts
1163 -> 1107 lines.
