+++ packages/store-core/src/handlers/index.ts
@@ -25,6 +25,11 @@ import {
   resolveSessionId,
 } from './_shared'
 import type { SessionPatch } from './_shared'
+// handleAuthOk folds the available-permission-modes list into its payload, so
+// it imports these back from ./permission (where the permission-mode handlers
+// now live — audit P2-3). No cycle: ./permission never imports ./index.
+import { handleAvailablePermissionModes } from './permission'
+import type { PermissionMode } from './permission'
 // ---------------------------------------------------------------------------
 // Shared helpers (parseStringField / parseRawStringField / parseEnumField /
 // resolveSessionId) and the SessionPatch type now live in ./_shared.ts
@@ -45,52 +50,12 @@ export function handleModelChanged(msg: Record<string, unknown>): { model: strin
 }
 
 // ---------------------------------------------------------------------------
-// permission_mode_changed
+// Permission-mode handlers (permission_mode_changed, available_permission_modes
+// — handleAvailablePermissionModes + PermissionMode, confirm_permission_mode)
+// live in ./permission.ts (audit P2-3 split), alongside the permission-request
+// handlers. Re-exported via the ./permission barrel line below. handleAuthOk
+// imports handleAvailablePermissionModes + PermissionMode back from there.
 // ---------------------------------------------------------------------------
-
-/** Extract the permission mode from a `permission_mode_changed` message. */
-export function handlePermissionModeChanged(msg: Record<string, unknown>): { mode: string | null } {
-  return { mode: parseStringField(msg, 'mode') }
-}
-
-// ---------------------------------------------------------------------------
-// available_permission_modes
-// ---------------------------------------------------------------------------
-
-export interface PermissionMode {
-  id: string
-  label: string
-  /**
-   * #4019: optional human-readable explainer for the mode (e.g. "Auto-approve
-   * every tool call without prompting"). The server's PERMISSION_MODES table
-   * exports a description for every mode; we keep the field optional here so
-   * older servers that pre-date the description plumbing still parse cleanly.
-   */
-  description?: string
-}
-
-/** Validate and extract permission modes from an `available_permission_modes` message. */
-export function handleAvailablePermissionModes(
-  msg: Record<string, unknown>,
-): PermissionMode[] | null {
-  if (!Array.isArray(msg.modes)) return null
-  return (msg.modes as unknown[])
-    .filter(
-      (m): m is { id: string; label: string; description?: unknown } =>
-        typeof m === 'object' &&
-        m !== null &&
-        typeof (m as { id: unknown }).id === 'string' &&
-        typeof (m as { label: unknown }).label === 'string',
-    )
-    .map((m) => {
-      const out: PermissionMode = { id: m.id, label: m.label }
-      // #4019: pass through description when present + string-typed. Non-
-      // strings (number, object, etc.) get dropped at the type boundary
-      // rather than poisoning the typed shape downstream consumers see.
-      if (typeof m.description === 'string') out.description = m.description
-      return out
-    })
-}
 
 // ---------------------------------------------------------------------------
 // Session-lifecycle handlers (session_updated, session_error, session_stopped,
@@ -100,30 +65,10 @@ export function handleAvailablePermissionModes(
 export * from './session-lifecycle'
 
 // ---------------------------------------------------------------------------
-// confirm_permission_mode
+// confirm_permission_mode (handleConfirmPermissionMode + PendingPermissionConfirm)
+// moved to ./permission.ts (audit P2-3 split). Re-exported via the ./permission
+// barrel line below.
 // ---------------------------------------------------------------------------
-
-export interface PendingPermissionConfirm {
-  mode: string
-  warning: string
-}
-
-/**
- * Extract the mode + warning text from a `confirm_permission_mode` message.
- *
- * Returns the pending-confirmation payload when the server included a valid
- * `mode` string, or null when the message is malformed (caller should leave
- * existing pending state alone in that case — matches both clients' prior
- * inline behavior).
- */
-export function handleConfirmPermissionMode(
-  msg: Record<string, unknown>,
-): PendingPermissionConfirm | null {
-  const mode = typeof msg.mode === 'string' ? msg.mode : null
-  if (!mode) return null
-  const warning = typeof msg.warning === 'string' ? msg.warning : 'Are you sure?'
-  return { mode, warning }
-}
 
 // ---------------------------------------------------------------------------
 // claude_ready
